+++ src/google-sheets/service/google-sheets.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { GoogleSheetsRepository } from "../domain/entities/google-sheets.repository";
+import { GoogleSheetsRepository } from "../domain/repository/google-sheets.repository";
 import { AddValue } from "src/lib/add-value.type";
 
 
@@ -21,6 +21,17 @@ export class GoogleSheetsService {
     try {
       const data = await this.googleSheetsRepository.getLasRowValue(range);
       return data;
+    } catch (error) {
+      this.googleSheetsRepository.failure(error);
+    }
+  }
+
+  async checkDate(date: string): Promise<boolean> {
+    try {
+      const data = await this.googleSheetsRepository.getDates();
+
+      const dateExists = data.some(row => row[0] === date);
+      return dateExists;
     } catch (error) {
       this.googleSheetsRepository.failure(error);
     }