+++ src/modules/dates/controller/dates.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Post, Body } from '@nestjs/common';
+import { Controller, Post, Body, Get, Query } from '@nestjs/common';
 import { DatesService } from '../service/dates.service';
 
 @Controller('dates')
@@ -10,6 +10,13 @@ export class DatesController {
     return this.datesService.createDate();
   }
 
+  @Get()
+  checkDate(
+    @Body('date') date: string
+  ): Promise<boolean> {
+    return this.datesService.checkDate(date);
+  }
+  
   @Post('/next-date')
   createNextDate(): Promise<string> {
     return this.datesService.createNextDate();
@@ -19,5 +26,6 @@ export class DatesController {
   createXDates(@Body('quantity') quantity: number): Promise<string> {
     return this.datesService.createXDates(quantity);
   }
+
 
 }