+++ src/modules/dates/dates.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { DatesService } from './service/dates.service';
 import { DatesController } from './controller/dates.controller';
-import { GoogleSheetsModule } from 'src/shared/google-sheets.module';
+import { GoogleSheetsModule } from 'src/google-sheets/google-sheets.module';
 import { GenerateDatetime } from './dateTime-build/generate-datetime';
 import { CreateDayUseCase } from './aplication/create-day.use-case';